SSE 端点从 sse.ts 迁移到 messages 路由，解决路由冲突
