+++ noj-core/tests/services/messages.test.ts
@@ -62,7 +62,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       assertEquals(typeof conv.id, "string");
       assertEquals(conv.user1_id, userA < userB ? userA : userB);
       assertEquals(conv.user2_id, userA < userB ? userB : userA);
@@ -81,8 +81,8 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv1 = await findOrCreateConversation(userA, userB);
-      const conv2 = await findOrCreateConversation(userA, userB);
+      const { conversation: conv1 } = await findOrCreateConversation(userA, userB);
+      const { conversation: conv2 } = await findOrCreateConversation(userA, userB);
       assertEquals(conv1.id, conv2.id);
     } finally {
       await cleanup(userA, userB);
@@ -137,7 +137,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       const msg = await sendMessage(userA, conv.id, "Hello!");
       assertEquals(msg.content, "Hello!");
       assertEquals(msg.sender_id, userA);
@@ -157,7 +157,7 @@ Deno.test({
     const userB = await createTestUser();
     const userC = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       await assertRejects(
         () => sendMessage(userC, conv.id, "Hi"),
         NotFoundError,
@@ -212,7 +212,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       await sendMessage(userA, conv.id, "msg1");
       await sendMessage(userB, conv.id, "msg2");
       const result = await listMessages(userA, conv.id, 1, 10);
@@ -233,7 +233,7 @@ Deno.test({
     const userB = await createTestUser();
     const userC = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       await assertRejects(
         () => listMessages(userC, conv.id, 1, 10),
         NotFoundError,
@@ -253,7 +253,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       const msg1 = await sendMessage(userA, conv.id, "test message");
 
       // 标记已读前 B 有 1 条未读
@@ -279,7 +279,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       await sendMessage(userA, conv.id, "msg1");
 
       const total = await getUnreadCount(userB);
@@ -300,7 +300,7 @@ Deno.test({
     const userA = await createTestUser();
     const userB = await createTestUser();
     try {
-      const conv = await findOrCreateConversation(userA, userB);
+      const { conversation: conv } = await findOrCreateConversation(userA, userB);
       const msg = await sendMessage(userA, conv.id, "delete me");
 
       // A 删除这条消息
